Crate crosvm_plugin

This module implements the dynamically loaded client library API used by a crosvm plugin, defined in crosvm.h. It implements the client half of the plugin protocol, which is defined in the protos::plugin module.

To implement the crosvm.h C API, each function and struct definition is repeated here, with concrete definitions for each struct. Most functions are thin shims to the underlying object oriented Rust implementation method. Most methods require a request over the crosvm connection, which is done by creating a MainRequest or VcpuRequest protobuf and sending it over the connection’s socket. Then, that socket is read for a MainResponse or VcpuResponse, which is translated to the appropriate return type for the C API.

Structs

anon_irqchip

anon_msi

crosvm

crosvm_hint

crosvm_hint_detail

crosvm_io_event

crosvm_irq_event

crosvm_irq_route

crosvm_memory

crosvm_net_config

crosvm_vcpu

crosvm_vcpu_event

crosvm_vcpu_reg_cache

Enums

Stat

Functions

crosvm_check_extension⚠

crosvm_connect⚠

crosvm_create_io_event⚠

crosvm_create_irq_event⚠

crosvm_create_memory⚠

crosvm_destroy_connection⚠

crosvm_destroy_io_event⚠

crosvm_destroy_irq_event⚠

crosvm_destroy_memory⚠

crosvm_enable_capability⚠

crosvm_get_clock⚠

crosvm_get_emulated_cpuid⚠

crosvm_get_hyperv_cpuid⚠

crosvm_get_ioapic_state⚠

crosvm_get_msr_index_list⚠

crosvm_get_pic_state⚠

crosvm_get_pit_state⚠

crosvm_get_shutdown_eventfd⚠

crosvm_get_supported_cpuid⚠

crosvm_get_vcpu⚠

crosvm_io_event_fd⚠

crosvm_irq_event_get_fd⚠

crosvm_irq_event_get_resample_fd⚠

crosvm_memory_get_dirty_log⚠

crosvm_net_get_config⚠

crosvm_new_connection⚠

crosvm_pause_vcpus⚠

crosvm_reserve_async_write_range⚠

crosvm_reserve_range⚠

crosvm_set_clock⚠

crosvm_set_hypercall_hint⚠

crosvm_set_identity_map_addr⚠

crosvm_set_ioapic_state⚠

crosvm_set_irq⚠

crosvm_set_irq_routing⚠

crosvm_set_pic_state⚠

crosvm_set_pit_state⚠

crosvm_start⚠

crosvm_vcpu_enable_capability⚠

crosvm_vcpu_get_debugregs⚠

crosvm_vcpu_get_fpu⚠

crosvm_vcpu_get_lapic_state⚠

crosvm_vcpu_get_mp_state⚠

crosvm_vcpu_get_msrs⚠

crosvm_vcpu_get_regs⚠

crosvm_vcpu_get_sregs⚠

crosvm_vcpu_get_vcpu_events⚠

crosvm_vcpu_get_xcrs⚠

crosvm_vcpu_resume⚠

crosvm_vcpu_set_cpuid⚠

crosvm_vcpu_set_debugregs⚠

crosvm_vcpu_set_fpu⚠

crosvm_vcpu_set_lapic_state⚠

crosvm_vcpu_set_mp_state⚠

crosvm_vcpu_set_msrs⚠

crosvm_vcpu_set_regs⚠

crosvm_vcpu_set_sregs⚠

crosvm_vcpu_set_vcpu_events⚠

crosvm_vcpu_set_xcrs⚠

crosvm_vcpu_wait⚠

Unions

anon_route